requirements.txt
fastapi
uvicorn[standard]

python-dotenv
mysql-connector-python

bcrypt
PyJWT

langchain
langchain-core
langchain-community
langchain-ollama

faiss-cpu

pypdf
docx2txt

tiktoken